move stores into a module for organization
